+++ WForest/src/Props/Text/FontSize.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using Serilog;
 using WForest.Exceptions;
@@ -39,9 +38,18 @@ namespace WForest.Props.Text
         {
             ApplicationDone = false;
             if (_size < 0) throw new ArgumentException("FontSize cannot be negative.");
+            
+            var tail = widget.Skip(1).OfType<Widgets.BuiltIn.Text>().ToList();
+            if (widget is Widgets.BuiltIn.Text text) text.FontSize = _size;
 
-            var ts = widget.OfType<Widgets.BuiltIn.Text>().ToList();
-            if (ts.Count > 0) foreach (var t in ts) t.FontSize = _size;
+            if (tail.Count > 0)
+            {
+                foreach (var t in tail)
+                {
+                    var alreadyHasFontFamilyProp = t.Props.SafeGetByProp<FontSize>().TryGetValue(out var l);
+                    if (!alreadyHasFontFamilyProp || l.Count == 0) t.FontSize = _size;
+                }
+            }
             else
                 Log.Warning(
                     "FontSize was applied to a widget that is not a Text nor has any Text in its sub-tree");